+++ actions/CostAnalyzerCsvExport.php
@@ -18,7 +18,7 @@ class CostAnalyzerCsvExport extends CController {
 	protected function checkInput(): bool {
 		return $this->validateInput([
 			'filter_groupids' => 'array',
-			'filter_host' => 'string',
+			'filter_hostids' => 'array',
 			'filter_status' => 'in -1,'.HOST_STATUS_MONITORED.','.HOST_STATUS_NOT_MONITORED,
 			'filter_show_maintenance' => 'in '.HOST_MAINTENANCE_STATUS_OFF.','.HOST_MAINTENANCE_STATUS_ON,
 			'filter_evaltype' => 'in '.TAG_EVAL_TYPE_AND_OR.','.TAG_EVAL_TYPE_OR,
@@ -43,7 +43,7 @@ class CostAnalyzerCsvExport extends CController {
 	}
 
 	private function hasFilterInput(): bool {
-		foreach (['filter_groupids', 'filter_host', 'filter_status', 'filter_show_maintenance', 'filter_tags'] as $field) {
+		foreach (['filter_groupids', 'filter_hostids', 'filter_status', 'filter_show_maintenance', 'filter_tags'] as $field) {
 			if ($this->hasInput($field)) {
 				return true;
 			}
@@ -55,7 +55,7 @@ class CostAnalyzerCsvExport extends CController {
 	private function getFilterInput(): array {
 		return [
 			'groupids' => $this->getInput('filter_groupids', []),
-			'host' => $this->getInput('filter_host', ''),
+			'hostids' => $this->getInput('filter_hostids', []),
 			'status' => $this->getInput('filter_status', -1),
 			'show_maintenance' => $this->getInput('filter_show_maintenance', HOST_MAINTENANCE_STATUS_ON),
 			'evaltype' => $this->getInput('filter_evaltype', TAG_EVAL_TYPE_AND_OR),